1040 - Averege 3 - Trocado double por float (5%)

+++ AverageTree/AverageTree/Program.cs
@@ -12,12 +12,12 @@ namespace AverageTree
             string score = Console.ReadLine();
             string[] vs = score.Split(' ');
 
-            double N1 = double.Parse(vs[0], CultureInfo.InvariantCulture);
-            double N2 = double.Parse(vs[1], CultureInfo.InvariantCulture);
-            double N3 = double.Parse(vs[2], CultureInfo.InvariantCulture);
-            double N4 = double.Parse(vs[3], CultureInfo.InvariantCulture);
+            float N1 = float.Parse(vs[0], CultureInfo.InvariantCulture);
+            float N2 = float.Parse(vs[1], CultureInfo.InvariantCulture);
+            float N3 = float.Parse(vs[2], CultureInfo.InvariantCulture);
+            float N4 = float.Parse(vs[3], CultureInfo.InvariantCulture);
 
-            double average = ((N1 * 2) + (N2 * 3) + (N3 * 4) + N4) / 10;
+            float average = ((N1 * 2) + (N2 * 3) + (N3 * 4) + N4) / 10;
 
             Console.WriteLine("Media: " + average.ToString("F1", CultureInfo.InvariantCulture));
 
@@ -29,7 +29,7 @@ namespace AverageTree
             {
                 Console.WriteLine("Aluno em exame.");
                 Console.Write("Nota do exame: ");
-                double examScore = double.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
+                float examScore = float.Parse(Console.ReadLine(), CultureInfo.InvariantCulture);
                 average = (average + examScore) / 2;
                 if (average >= 5)
                 {
